app/requirements.txt
Pillow==9.1.1
fastapi==0.78.0
pydantic==1.9.1
numpy==1.22.4
uvicorn==0.17.6
boto3==1.24.9
transformers==4.30.2
python-multipart==0.0.6
torch==2.0.1
torchvision==0.15.2
faiss-cpu==1.7.4
scikit-learn==1.2.2
pandas==2.0.2
requests==2.31.0
aiofiles==23.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1